Migrate to FileListItem from FileListItemV3
The UI package contained two list item components, and to streamline the component structure, the old component FileListItem has been replaced with its version 3 counterpart, FileListItemV3, across all relevant components.

Summary:

- Migrated from the outdated FileListItem to FileListItemV3 in all related components.
- Deleted the legacy FileListItem.svelte component to avoid confusion and maintain consistency.
- Renamed FileListItemV3.svelte to FileListItem.svelte, consolidating component references to enhance readability and manageability.
- Updated desktop app components to reflect the new nomenclature and ensure seamless functionality.
- Adjusted the story files to align with the new component structure while removing obsolete story files to maintain a clean repository.

These changes were necessary to remove redundancy and streamline component usage within the UI package, ensuring a consistent and efficient component architecture. The refactoring reflects our commitment to maintaining a modern and simplified codebase.
