+++ apps/desktop/src/components/FileListItemWrapper.svelte
@@ -14,7 +14,7 @@
 	import { TestId } from '$lib/testing/testIds';
 	import { computeChangeStatus } from '$lib/utils/fileStatus';
 	import { inject } from '@gitbutler/shared/context';
-	import FileListItemV3 from '@gitbutler/ui/file/FileListItemV3.svelte';
+	import FileListItem from '@gitbutler/ui/file/FileListItem.svelte';
 	import FileViewHeader from '@gitbutler/ui/file/FileViewHeader.svelte';
 	import { stickyHeader } from '@gitbutler/ui/utils/stickyHeader';
 	import type { ConflictEntriesObj } from '$lib/files/conflicts';
@@ -174,7 +174,7 @@
 			oncloseclick={onCloseClick}
 		/>
 	{:else}
-		<FileListItemV3
+		<FileListItem
 			id={key({ ...selectionId, path: change.path })}
 			filePath={change.path}
 			fileStatus={computeChangeStatus(change)}
