+++ packages/ui/src/lib/file/FileListItem.svelte
@@ -1,11 +1,13 @@
 <script lang="ts">
 	import Badge from '$lib/Badge.svelte';
+	import Button from '$lib/Button.svelte';
 	import Checkbox from '$lib/Checkbox.svelte';
 	import Icon from '$lib/Icon.svelte';
 	import Tooltip from '$lib/Tooltip.svelte';
-	import FileIcon from '$lib/file/FileIcon.svelte';
+	import ExecutableLabel from '$lib/file/ExecutableLabel.svelte';
+	import FileIndent from '$lib/file/FileIndent.svelte';
+	import FileName from '$lib/file/FileName.svelte';
 	import FileStatusBadge from '$lib/file/FileStatusBadge.svelte';
-	import { splitFilePath } from '$lib/utils/filePath';
 	import type { FileStatus } from '$lib/file/types';
 
 	interface Props {
@@ -13,23 +15,32 @@
 		id?: string;
 		filePath: string;
 		fileStatus?: FileStatus;
+		fileStatusTooltip?: string;
 		fileStatusStyle?: 'dot' | 'full';
 		draggable?: boolean;
 		selected?: boolean;
+		focused?: boolean;
 		clickable?: boolean;
 		showCheckbox?: boolean;
+		listMode?: 'list' | 'tree';
+		depth?: number;
 		checked?: boolean;
 		indeterminate?: boolean;
 		conflicted?: boolean;
 		conflictHint?: string;
 		locked?: boolean;
 		lockText?: string;
+		active?: boolean;
+		isLast?: boolean;
+		executable?: boolean;
+		oncheckclick?: (e: MouseEvent) => void;
 		oncheck?: (
 			e: Event & {
 				currentTarget: EventTarget & HTMLInputElement;
 			}
 		) => void;
 		onclick?: (e: MouseEvent) => void;
+		ondblclick?: (e: MouseEvent) => void;
 		onresolveclick?: (e: MouseEvent) => void;
 		onkeydown?: (e: KeyboardEvent) => void;
 		oncontextmenu?: (e: MouseEvent) => void;
@@ -40,9 +51,12 @@
 		id,
 		filePath,
 		fileStatus,
+		fileStatusTooltip,
+		isLast,
 		fileStatusStyle = 'dot',
 		draggable = false,
 		selected = false,
+		focused = false,
 		clickable = true,
 		showCheckbox = false,
 		checked = $bindable(),
@@ -51,14 +65,20 @@
 		conflictHint,
 		locked,
 		lockText,
+		active,
+		listMode = 'list',
+		depth,
+		executable,
 		oncheck,
+		oncheckclick,
 		onclick,
+		ondblclick,
 		onresolveclick,
 		onkeydown,
 		oncontextmenu
 	}: Props = $props();
 
-	const fileInfo = $derived(splitFilePath(filePath));
+	const showIndent = $derived(depth && depth > 0);
 </script>
 
 <div
@@ -66,14 +86,18 @@
 	data-locked={locked}
 	data-file-id={id}
 	class="file-list-item"
-	class:selected-draggable={selected}
+	class:selected
+	class:active
 	class:clickable
+	class:focused
 	class:draggable
-	class:open
+	class:last={isLast}
+	class:list-mode={listMode === 'list'}
 	aria-selected={selected}
 	role="option"
 	tabindex="-1"
 	{onclick}
+	{ondblclick}
 	{onkeydown}
 	oncontextmenu={(e) => {
 		if (oncontextmenu) {
@@ -83,25 +107,27 @@
 		}
 	}}
 >
-	{#if showCheckbox}
-		<Checkbox small {checked} {indeterminate} onchange={oncheck} />
-	{/if}
-	<div class="info">
-		<FileIcon fileName={fileInfo.filename} />
-		<span class="text-12 text-semibold name truncate">
-			{fileInfo.filename}
-		</span>
-
-		<div class="path-container">
-			<Tooltip text={filePath} delay={1200}>
-				<span class="text-12 path truncate">
-					{fileInfo.path}
-				</span>
-			</Tooltip>
+	{#if draggable && !showCheckbox}
+		<div class="draggable-handle">
+			<Icon name="draggable-narrow" />
 		</div>
-	</div>
+	{/if}
 
-	<div class="details">
+	{#if showIndent || showCheckbox}
+		<div class="file-list-item__indicators">
+			{#if showIndent}
+				<FileIndent {depth} />
+			{/if}
+
+			{#if showCheckbox}
+				<Checkbox small {checked} {indeterminate} onchange={oncheck} onclick={oncheckclick} />
+			{/if}
+		</div>
+	{/if}
+
+	<FileName {filePath} hideFilePath={listMode === 'tree'} />
+
+	<div class="file-list-item__details">
 		{#if locked}
 			<Tooltip text={lockText}>
 				<div class="locked">
@@ -116,17 +142,19 @@
 					<Badge style="success">Resolved</Badge>
 				</Tooltip>
 			{:else}
-				<button
+				<Button
 					type="button"
+					kind="outline"
 					class="mark-resolved-btn"
+					size="tag"
 					onclick={(e) => {
 						e.stopPropagation();
 						onresolveclick?.(e);
 					}}
+					icon="tick-small"
 				>
-					<span class="text-11 text-semibold">Mark as resolved</span>
-					<Icon name="tick-small" opacity={0.5} />
-				</button>
+					Mark as resolved
+				</Button>
 			{/if}
 		{/if}
 
@@ -138,14 +166,12 @@
 			</Tooltip>
 		{/if}
 
-		{#if fileStatus}
-			<FileStatusBadge status={fileStatus} style={fileStatusStyle} />
+		{#if executable}
+			<ExecutableLabel />
 		{/if}
 
-		{#if draggable}
-			<div class="draggable-handle">
-				<Icon name="draggable-narrow" />
-			</div>
+		{#if fileStatus}
+			<FileStatusBadge tooltip={fileStatusTooltip} status={fileStatus} style={fileStatusStyle} />
 		{/if}
 	</div>
 </div>
@@ -153,127 +179,81 @@
 <style lang="postcss">
 	.file-list-item {
 		display: flex;
+		position: relative;
 		align-items: center;
 		height: 32px;
-		padding: 6px 14px;
+		padding: 0 10px 0 14px;
 		overflow: hidden;
-		gap: 10px;
-		border-bottom: none;
+		gap: 8px;
 		outline: none;
-		background: transparent;
+		background-color: var(--clr-bg-1);
 		text-align: left;
 		user-select: none;
 
-		&:last-child {
-			border-bottom: none;
+		& :global(.mark-resolved-btn) {
+			margin: 0 4px;
 		}
 
-		&:not(:last-child) {
+		&.clickable {
+			cursor: pointer;
+
+			&:not(.selected):hover {
+				background-color: var(--clr-bg-1-muted);
+			}
+		}
+
+		&.selected {
+			background-color: var(--clr-selected-not-in-focus-bg);
+		}
+
+		&.active.selected {
+			background-color: var(--clr-selected-in-focus-bg);
+		}
+
+		&.list-mode:not(.last) {
 			border-bottom: 1px solid var(--clr-border-3);
 		}
-	}
 
-	.file-list-item.clickable {
-		cursor: pointer;
-
-		&:not(.selected-draggable):hover {
-			background-color: var(--clr-bg-1-muted);
-		}
-	}
-
-	.draggable {
-		&:hover {
-			& .draggable-handle {
-				opacity: 1;
+		&.draggable {
+			&:hover {
+				& .draggable-handle {
+					opacity: 1;
+				}
 			}
 		}
 	}
 
-	.draggable-handle {
+	.file-list-item__indicators {
 		display: flex;
 		align-items: center;
-		justify-content: center;
-		width: 4px;
-		margin-right: -8px;
-		color: var(--clr-text-3);
-		opacity: 0;
-		transition:
-			width var(--transition-fast),
-			opacity var(--transition-fast);
-	}
-
-	.mark-resolved-btn {
-		display: flex;
-		align-items: center;
-		margin: 0 2px;
-		padding: 3px 6px 3px 6px;
-		gap: 4px;
-		border: 1px solid var(--clr-border-2);
-		border-radius: var(--radius-m);
-		white-space: nowrap;
-		transition:
-			background-color var(--transition-fast),
-			border-color var(--transition-fast);
-
-		&:hover {
-			background-color: var(--clr-bg-1);
-		}
-	}
-
-	.info {
-		display: flex;
-		flex-grow: 1;
-		flex-shrink: 1;
-		align-items: center;
-		min-width: 32px;
-		overflow: hidden;
 		gap: 6px;
 	}
 
-	.name {
-		flex-grow: 0;
-		flex-shrink: 1;
-		min-width: 40px;
-		color: var(--clt-text-1);
-		pointer-events: none;
-	}
-
-	.path-container {
+	.draggable-handle {
 		display: flex;
-		flex-grow: 1;
-		flex-shrink: 0;
-		flex-basis: 0px;
-		justify-content: flex-start;
-		min-width: 50px;
-		overflow: hidden;
-		text-align: left;
-	}
-
-	.path {
-		display: inline-block;
-		max-width: 100%;
-		color: var(--clt-text-1);
-		line-height: 120%;
-		text-align: left;
-		opacity: 0.3;
+		position: absolute;
+		left: 0;
+		align-items: center;
+		justify-content: center;
+		height: 24px;
+		color: var(--clr-text-3);
+		opacity: 0;
 		transition: opacity var(--transition-fast);
 	}
 
-	.details {
+	.file-list-item__details {
 		display: flex;
+		flex-grow: 1;
 		align-items: center;
 		gap: 6px;
-	}
 
-	.details .locked {
-		display: flex;
-	}
+		& .locked {
+			display: flex;
+		}
 
-	.details .conflicted {
-		display: flex;
-	}
-
-	.selected-draggable {
-		background-color: var(--clr-theme-pop-bg-muted) !important;
+		& .conflicted {
+			display: flex;
+			margin-right: -2px;
+		}
 	}
 </style>
